Move amazon to platforms folder
