--- src/amazon/amazon_settings.rs
@@ -1,7 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct AmazonSettings {
-    pub enabled: bool,
-    pub launcher_location: Option<String>,
-}
