--- src/amazon/mod.rs
@@ -1,8 +1,0 @@
-pub mod amazon_platform;
-pub use amazon_platform::*;
-
-pub mod amazon_game;
-pub use amazon_game::*;
-
-pub mod amazon_settings;
-pub use amazon_settings::*;
