+++ src/platforms/amazon/amazon_platform.rs
@@ -1,8 +1,7 @@
 use std::path::{Path, PathBuf};
-
 use sqlite::State;
-
-use super::{AmazonGame, AmazonSettings};
+use serde::{Deserialize, Serialize};
+use steam_shortcuts_util::{shortcut::ShortcutOwned, Shortcut};
 
 #[derive(Clone)]
 pub struct AmazonPlatform {
@@ -70,3 +69,41 @@ impl AmazonPlatform {
         ui.checkbox(&mut self.settings.enabled, "Import from Amazon");
     }
 }
+
+
+#[derive(Debug, Clone)]
+pub struct AmazonGame {
+    pub title: String,
+    pub id: String,
+    pub launcher_path: PathBuf,
+}
+
+impl From<AmazonGame> for ShortcutOwned {
+    fn from(game: AmazonGame) -> Self {
+        let launch = format!("amazon-games://play/{}", game.id);
+        let exe = game.launcher_path.to_string_lossy().to_string();
+        let start_dir = game
+            .launcher_path
+            .parent()
+            .unwrap_or_else(|| Path::new(""))
+            .to_string_lossy()
+            .to_string();
+        Shortcut::new(
+            "0",
+            game.title.as_str(),
+            exe.as_str(),
+            start_dir.as_str(),
+            "",
+            "",
+            launch.as_str(),
+        )
+        .to_owned()
+    }
+}
+
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct AmazonSettings {
+    pub enabled: bool,
+    pub launcher_location: Option<String>,
+}
+
