+++ src/platforms/amazon/mod.rs
@@ -1,0 +1,5 @@
+mod amazon_platform;
+
+
+pub use amazon_platform::AmazonSettings;
+pub use amazon_platform::AmazonPlatform;
