+++ src/platforms/mod.rs
@@ -1,6 +1,8 @@
 mod platform;
 mod egs;
+mod amazon;
 
 pub use platform::*;
 
+pub use amazon::AmazonSettings;
 pub use egs::EpicGamesLauncherSettings;
